Fix frontend modal not being openable twice
